libstdc++: Fix -Wsign-compare warnings in bits/hashtable_policy.h
libstdc++-v3/ChangeLog:

	* include/bits/hashtable_policy.h (_Local_iterator_base): Fix
	-Wsign-compare warnings.

+++ libstdc++-v3/include/bits/hashtable_policy.h
@@ -1302,12 +1302,12 @@ namespace __detail
       _Local_iterator_base&
       operator=(const _Local_iterator_base& __iter)
       {
-	if (_M_bucket_count != -1)
+	if (_M_bucket_count != size_t(-1))
 	  _M_destroy();
 	this->_M_cur = __iter._M_cur;
 	_M_bucket = __iter._M_bucket;
 	_M_bucket_count = __iter._M_bucket_count;
-	if (_M_bucket_count != -1)
+	if (_M_bucket_count != size_t(-1))
 	  _M_init(*__iter._M_h());
 	return *this;
       }
